windows: resolve 8.3 short paths before querying disk size
GetFileInformationByHandleEx fails with an invalid-handle error when the
path contains 8.3 short names (ex.: C:\Users\RUNNER~1), which the CI temp
dir on windows produces. Resolve the long path form first via
GetLongPathName.

Moves the existing resolveLongPath helper out of the vhdx test into a
shared windows util so check_files can use it as well.
